Organizing file structure and styling menu
