+++ Lambda-Notes/src/components/Menu/Menu.css
@@ -1,0 +1,20 @@
+.Menu {
+  display: flex;
+  flex-flow: column;
+  padding-left: 10px;
+  width: 30%;
+  height: 50%;
+  background-color: #d7d7d7;
+  color: #414242;
+}
+
+.menu__button {
+  background-color: #24b8bd;
+  border: 1px black solid;
+  height: 35px;
+  width: 80%;
+  color: white;
+  margin-top: 10px;
+  font-weight: bold;
+  cursor: pointer;
+}
