+++ Lambda-Notes/src/index.css
@@ -1,3 +1,5 @@
+@import './components/Menu/Menu.css';
+
 body {
   margin: 0;
   padding: 0;
